refactor: move pack logic to own file
