+++ packages/express/src/app.ts
@@ -20,6 +20,7 @@ import {
   CourseCreationQueue,
   initCourseDBDesignDocInsert,
 } from './client-requests/course-requests.js';
+import { packCourse } from './client-requests/pack-requests.js';
 import { requestIsAuthenticated } from './couchdb/authentication.js';
 import CouchDB, {
   useOrCreateCourseDB,
@@ -37,7 +38,6 @@ logger.info(`Express app running version: ${ENV.VERSION}`);
 
 const port = 3000;
 import { classroomDbDesignDoc } from './design-docs.js';
-import PouchDb from 'pouchdb';
 const app = express();
 
 app.use(cookieParser());
@@ -171,112 +171,11 @@ async function postHandler(
           res.json(e);
         });
     } else if (body.type === RequestEnum.PACK_COURSE) {
-      logger.info(`Starting PACK_COURSE for ${body.courseId}...`);
-      
-      try {
-        const startTime = Date.now();
-        
-        // Use CouchDBToStaticPacker directly from db package
-        const { CouchDBToStaticPacker } = await import('@vue-skuilder/db');
-        
-        // Create database connection URL
-        const dbUrl = `${ENV.COUCHDB_PROTOCOL}://${ENV.COUCHDB_ADMIN}:${ENV.COUCHDB_PASSWORD}@${ENV.COUCHDB_SERVER}`;
-        const dbName = `coursedb-${body.courseId}`;
-        
-        // Determine output path - for studio mode, use a more appropriate directory
-        const outputPath = body.outputPath || (ENV.NODE_ENV === 'studio' ? 
-          '/tmp/skuilder-studio-output' : 
-          process.cwd());
-        
-        logger.info(`Packing course ${body.courseId} from ${dbName} to ${outputPath}`);
-        
-        // Create course database connection
-        const courseDbUrl = `${dbUrl}/${dbName}`;
-        
-        // Initialize packer and perform pack operation with file writing
-        const packer = new CouchDBToStaticPacker();
-        
-        // For Express, we create a simple FileSystemAdapter using dynamic imports
-        const createFsAdapter = async () => {
-          const fsExtra = await import('fs-extra');
-          const pathModule = await import('path');
-          
-          // Access the default export for fs-extra
-          const fs = fsExtra.default || fsExtra;
-          const path = pathModule.default || pathModule;
-          
-          return {
-            async readFile(filePath: string): Promise<string> {
-              return await fs.readFile(filePath, 'utf8');
-            },
-            async readBinary(filePath: string): Promise<Buffer> {
-              return await fs.readFile(filePath);
-            },
-            async exists(filePath: string): Promise<boolean> {
-              try {
-                await fs.access(filePath);
-                return true;
-              } catch {
-                return false;
-              }
-            },
-            async stat(filePath: string): Promise<any> {
-              const stats = await fs.stat(filePath);
-              return {
-                isDirectory: () => stats.isDirectory(),
-                isFile: () => stats.isFile(),
-                size: stats.size
-              };
-            },
-            async writeFile(filePath: string, data: string | Buffer): Promise<void> {
-              await fs.writeFile(filePath, data);
-            },
-            async writeJson(filePath: string, data: any, options?: { spaces?: number }): Promise<void> {
-              await fs.writeJson(filePath, data, options);
-            },
-            async ensureDir(dirPath: string): Promise<void> {
-              await fs.ensureDir(dirPath);
-            },
-            joinPath(...segments: string[]): string {
-              return path.join(...segments);
-            },
-            dirname(filePath: string): string {
-              return path.dirname(filePath);
-            },
-            isAbsolute(filePath: string): boolean {
-              return path.isAbsolute(filePath);
-            }
-          };
-        };
-        
-        const fsAdapter = await createFsAdapter();
-        const packResult = await packer.packCourseToFiles(new PouchDb(courseDbUrl), body.courseId, outputPath, fsAdapter);
-        
-        const duration = Date.now() - startTime;
-        
-        const response = {
-          status: 'ok' as const,
-          ok: true,
-          packedFiles: [], // No longer relevant since we're writing to files
-          outputPath: outputPath,
-          attachmentsFound: packResult.attachmentsFound,
-          filesWritten: packResult.filesWritten,
-          totalFiles: packResult.filesWritten, // Updated to reflect actual files written
-          duration: duration
-        };
-        
-        logger.info(`Pack completed in ${duration}ms. Attachments: ${response.attachmentsFound}, Files written: ${response.filesWritten}`);
-        res.json(response);
-        
-      } catch (error) {
-        logger.error('Pack operation failed:', error);
-        const response = {
-          status: 'error' as const,
-          ok: false,
-          errorText: error instanceof Error ? error.message : 'Pack operation failed'
-        };
-        res.status(500).json(response);
-      }
+      body.response = await packCourse({
+        courseId: body.courseId,
+        outputPath: body.outputPath
+      });
+      res.json(body.response);
     }
   } else {
     logger.info(`\tREQUEST UNAUTHORIZED!`);
